pyproject.toml
[tool.poetry]
name = "casemaster"
version = "0.1.0"
description = ""
authors = ["zero <dukepandme@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.13"
django = "3.2"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
